Refactor backdrop to be its own component rather than part of modal
