+++ src/components/Backdrop/animations.ts
@@ -1,0 +1,10 @@
+import { Variants } from "framer-motion";
+
+export const variants: Variants = {
+  hidden: {
+    opacity: 0,
+  },
+  active: {
+    opacity: 1,
+  }
+};
