+++ src/components/Backdrop/types.ts
@@ -1,0 +1,5 @@
+import { ReactElement } from "react";
+
+export interface BackdropProps {
+  children: ReactElement;
+}
